refactor(inference): remove nxd submodule
