+++ optimum/neuron/models/inference/auto_models.py
@@ -22,12 +22,12 @@ automatically by the neuron factory classes such as NeuronModelForCausalLM.
 import os
 
 from ..auto_model import register_neuron_model
-from .nxd.granite.modeling_granite import GraniteNxDModelForCausalLM
-from .nxd.llama.modeling_llama import LlamaNxDModelForCausalLM
-from .nxd.mixtral.modeling_mixtral import MixtralNxDModelForCausalLM
-from .nxd.phi3.modeling_phi3 import Phi3NxDModelForCausalLM
-from .nxd.qwen2.modeling_qwen2 import Qwen2NxDModelForCausalLM
-from .nxd.qwen3.modeling_qwen3 import Qwen3NxDModelForCausalLM
+from .granite.modeling_granite import GraniteNxDModelForCausalLM
+from .llama.modeling_llama import LlamaNxDModelForCausalLM
+from .mixtral.modeling_mixtral import MixtralNxDModelForCausalLM
+from .phi3.modeling_phi3 import Phi3NxDModelForCausalLM
+from .qwen2.modeling_qwen2 import Qwen2NxDModelForCausalLM
+from .qwen3.modeling_qwen3 import Qwen3NxDModelForCausalLM
 
 
 prioritize_hlo_backend = os.environ.get("OPTIMUM_NEURON_PRIORITIZE_HLO_BACKEND", "0") == "1"
