+++ optimum/neuron/models/neuron_config.py
@@ -14,5 +14,5 @@
 # limitations under the License.
 # This file is used by the NeuronConfig create method to register the model configurations
 
-from .inference.nxd.backend.config import NxDNeuronConfig  # noqa F401
+from .inference.backend.config import NxDNeuronConfig  # noqa F401
 from .training.config import TrainingNeuronConfig  # noqa F401
